Modify tables and endpoints

+++ Models/SparePart.cs
@@ -1,8 +1,8 @@
 namespace BikeDoctor.Models;
 
 public class SparePart
-{    
+{
     public string NameSparePart { get; set; }
     public string DetailSparePart { get; set; }
-    public int Price { get; set; }
+    public decimal Price { get; set; }
 }
